REF: share more methods in ExtensionIndex  (#37970)
