+++ pandas/core/indexes/category.py
@@ -6,7 +6,6 @@ import numpy as np
 from pandas._config import get_option
 
 from pandas._libs import index as libindex
-from pandas._libs.hashtable import duplicated_int64
 from pandas._libs.lib import no_default
 from pandas._typing import ArrayLike, Label
 from pandas.util._decorators import Appender, cache_readonly, doc
@@ -14,7 +13,6 @@ from pandas.util._decorators import Appender, cache_readonly, doc
 from pandas.core.dtypes.common import (
     ensure_platform_int,
     is_categorical_dtype,
-    is_list_like,
     is_scalar,
 )
 from pandas.core.dtypes.dtypes import CategoricalDtype
@@ -226,9 +224,14 @@ class CategoricalIndex(NDArrayBackedExtensionIndex, accessor.PandasDelegate):
 
     # --------------------------------------------------------------------
 
+    # error: Argument 1 of "_shallow_copy" is incompatible with supertype
+    #  "ExtensionIndex"; supertype defines the argument type as
+    #  "Optional[ExtensionArray]"  [override]
     @doc(Index._shallow_copy)
-    def _shallow_copy(
-        self, values: Optional[Categorical] = None, name: Label = no_default
+    def _shallow_copy(  # type:ignore[override]
+        self,
+        values: Optional[Categorical] = None,
+        name: Label = no_default,
     ):
         name = self.name if name is no_default else name
 
@@ -247,6 +250,10 @@ class CategoricalIndex(NDArrayBackedExtensionIndex, accessor.PandasDelegate):
         provide a comparison between the dtype of self and other (coercing if
         needed)
 
+        Parameters
+        ----------
+        other : Index
+
         Returns
         -------
         Categorical
@@ -263,8 +270,6 @@ class CategoricalIndex(NDArrayBackedExtensionIndex, accessor.PandasDelegate):
                 )
         else:
             values = other
-            if not is_list_like(values):
-                values = [values]
 
             cat = Categorical(other, dtype=self.dtype)
             other = CategoricalIndex(cat)
@@ -358,11 +363,6 @@ class CategoricalIndex(NDArrayBackedExtensionIndex, accessor.PandasDelegate):
         """ return the underlying data, which is a Categorical """
         return self._data
 
-    @property
-    def _has_complex_internals(self) -> bool:
-        # used to avoid libreduction code paths, which raise or require conversion
-        return True
-
     @doc(Index.__contains__)
     def __contains__(self, key: Any) -> bool:
         # if key is a NaN, check if any NaN is in self.
@@ -398,11 +398,6 @@ class CategoricalIndex(NDArrayBackedExtensionIndex, accessor.PandasDelegate):
         # Use _simple_new instead of _shallow_copy to ensure we keep dtype
         #  of result, not self.
         return type(self)._simple_new(result, name=self.name)
-
-    @doc(Index.duplicated)
-    def duplicated(self, keep="first"):
-        codes = self.codes.astype("i8")
-        return duplicated_int64(codes, keep)
 
     def _to_safe_for_reshape(self):
         """ convert to object if we are a categorical """
@@ -482,7 +477,7 @@ class CategoricalIndex(NDArrayBackedExtensionIndex, accessor.PandasDelegate):
         new_target = np.asarray(new_target)
         if is_categorical_dtype(target):
             new_target = Categorical(new_target, dtype=target.dtype)
-            new_target = target._shallow_copy(new_target, name=self.name)
+            new_target = type(self)._simple_new(new_target, name=self.name)
         else:
             new_target = Index(new_target, name=self.name)
 
@@ -506,7 +501,7 @@ class CategoricalIndex(NDArrayBackedExtensionIndex, accessor.PandasDelegate):
             # .reindex returns normal Index. Revert to CategoricalIndex if
             # all targets are included in my categories
             new_target = Categorical(new_target, dtype=self.dtype)
-            new_target = self._shallow_copy(new_target)
+            new_target = type(self)._simple_new(new_target, name=self.name)
 
         return new_target, indexer, new_indexer
