+++ pandas/core/indexes/datetimes.py
@@ -20,11 +20,8 @@ from pandas.util._decorators import cache_readonly, doc
 
 from pandas.core.dtypes.common import (
     DT64NS_DTYPE,
-    is_datetime64_any_dtype,
     is_datetime64_dtype,
     is_datetime64tz_dtype,
-    is_float,
-    is_integer,
     is_scalar,
 )
 from pandas.core.dtypes.missing import is_valid_nat_for_dtype
@@ -354,8 +351,6 @@ class DatetimeIndex(DatetimeTimedeltaMixin):
         """
         Can we compare values of the given dtype to our own?
         """
-        if not is_datetime64_any_dtype(dtype):
-            return False
         if self.tz is not None:
             # If we have tz, we can compare to tzaware
             return is_datetime64tz_dtype(dtype)
@@ -720,9 +715,6 @@ class DatetimeIndex(DatetimeTimedeltaMixin):
         """
         assert kind in ["loc", "getitem", None]
 
-        if is_float(label) or isinstance(label, time) or is_integer(label):
-            self._invalid_indexer("slice", label)
-
         if isinstance(label, str):
             freq = getattr(self, "freqstr", getattr(self, "inferred_freq", None))
             parsed, reso = parsing.parse_time_string(label, freq)
@@ -739,6 +731,9 @@ class DatetimeIndex(DatetimeTimedeltaMixin):
             return lower if side == "left" else upper
         elif isinstance(label, (self._data._recognized_scalars, date)):
             self._deprecate_mismatched_indexing(label)
+        else:
+            self._invalid_indexer("slice", label)
+
         return self._maybe_cast_for_get_loc(label)
 
     def _get_string_slice(self, key: str):
