+++ pandas/core/indexes/extension.py
@@ -1,10 +1,12 @@
 """
 Shared methods for Index subclasses backed by ExtensionArray.
 """
-from typing import List, TypeVar
+from typing import List, Optional, TypeVar
 
 import numpy as np
 
+from pandas._libs import lib
+from pandas._typing import Label
 from pandas.compat.numpy import function as nv
 from pandas.errors import AbstractMethodError
 from pandas.util._decorators import cache_readonly, doc
@@ -211,6 +213,24 @@ class ExtensionIndex(Index):
     __le__ = _make_wrapped_comparison_op("__le__")
     __ge__ = _make_wrapped_comparison_op("__ge__")
 
+    @doc(Index._shallow_copy)
+    def _shallow_copy(
+        self, values: Optional[ExtensionArray] = None, name: Label = lib.no_default
+    ):
+        name = self.name if name is lib.no_default else name
+
+        if values is not None:
+            return self._simple_new(values, name=name)
+
+        result = self._simple_new(self._data, name=name)
+        result._cache = self._cache
+        return result
+
+    @property
+    def _has_complex_internals(self) -> bool:
+        # used to avoid libreduction code paths, which raise or require conversion
+        return True
+
     # ---------------------------------------------------------------------
     # NDarray-Like Methods
 
@@ -251,7 +271,7 @@ class ExtensionIndex(Index):
     def repeat(self, repeats, axis=None):
         nv.validate_repeat(tuple(), dict(axis=axis))
         result = self._data.repeat(repeats, axis=axis)
-        return self._shallow_copy(result)
+        return type(self)._simple_new(result, name=self.name)
 
     def insert(self, loc: int, item):
         # ExtensionIndex subclasses must override Index.insert
